requirements.txt
Django==5.2.3
# psycopg2-binary==2.9.10
# psycopg[binary]>=3.2.0
python-decouple==3.8
django-crispy-forms==2.4
django-tables2==2.7.5
django-filter==23.3
reportlab==4.4.1
openpyxl==3.1.5
requests==2.32.4
#Pillow==11.2.1
chardet==5.2.0
et-xmlfile==2.0.0
python-dateutil==2.8.2

# Additional packages for production
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1

# SMS Services
twilio>=8.5.0

# Forms and UI Enhancement
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.4.0

# Authentication and Security
django-guardian>=2.4.0
django-phonenumber-field>=7.1.0
phonenumbers>=8.13.0

# Utilities
pytz>=2023.3
arrow>=1.2.0
numpy>=1.24.0

# Development packages
django-debug-toolbar==4.2.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2

# Monitoring and Health Checks
django-health-check>=3.17.0
sentry-sdk>=1.29.0

# Background Tasks
django-rq>=2.8.0
rq>=1.15.0
